fix scope by using a persistent structure
this works by cloning the scope, and thereby the internal values store. Instead of the Scope and Parent being stored as a reference, we store the values in the hashmap as references, which ensures that any values defined in the scope the parent was cloned from after the clone will not be added to the new scope, but ensures that any changes to the cloned references will be reflected.